Refactor: move HasChannel over to the interface mixins module
