+++ pyiron_workflow/io.py
@@ -18,7 +18,7 @@ from pyiron_workflow.channels import (
     InputSignal,
     OutputSignal,
 )
-from pyiron_workflow.has_channel import HasChannel
+from pyiron_workflow.has_interface_mixins import HasChannel
 from pyiron_workflow.has_to_dict import HasToDict
 from pyiron_workflow.snippets.logger import logger
 from pyiron_workflow.snippets.dotdict import DotDict
